some quickfix

+++ src/game/entities/creatures/Creature.java
@@ -38,14 +38,14 @@ public abstract class Creature extends Entity{
 			if(!collisionWithTile(tx, (int)(y + bounds.y)/Tile.TILEHEIGHT) && !collisionWithTile(tx, (int)(y + bounds.y + bounds.height)/ Tile.TILEHEIGHT)) {
 				x += xMove;
 			}else {
-				x = tx * Tile.TILEWIDTH - bounds.x - bounds.width;
+				x = (float) tx * Tile.TILEWIDTH - bounds.x - bounds.width;
 			}
 		}else if(xMove < 0) {//left
 			int tx = (int)(x + xMove + bounds.x ) / Tile.TILEWIDTH;
 			if(!collisionWithTile(tx, (int)(y + bounds.y)/Tile.TILEHEIGHT) && !collisionWithTile(tx, (int)(y + bounds.y + bounds.height)/ Tile.TILEHEIGHT)) {
 				x += xMove;
 			}else {
-				x = tx *Tile.TILEWIDTH + Tile.TILEWIDTH - bounds.x;
+				x = (float) tx *Tile.TILEWIDTH + Tile.TILEWIDTH - bounds.x;
 			}
 		}
 	}
@@ -57,7 +57,7 @@ public abstract class Creature extends Entity{
 					!collisionWithTile((int)(x + bounds.x + bounds.width)/Tile.TILEWIDTH, ty)) {
 				y += yMove;
 			}else {
-				y = ty * Tile.TILEHEIGHT + Tile.TILEHEIGHT - bounds.y;
+				y = (float) ty * Tile.TILEHEIGHT + Tile.TILEHEIGHT - bounds.y;
 			}
 		}else if(yMove > 0) {//down
 			int ty =(int)( y + yMove + bounds.y + bounds.height)/Tile.TILEHEIGHT;
@@ -65,7 +65,7 @@ public abstract class Creature extends Entity{
 					!collisionWithTile((int)(x + bounds.x + bounds.width)/Tile.TILEWIDTH, ty)) {
 				y += yMove;
 			}else {
-				y = ty * Tile.TILEHEIGHT - bounds.y - bounds.height ;
+				y = (float) ty * Tile.TILEHEIGHT - bounds.y - bounds.height ;
 			}
 		}
 	}
